WIP: moving the category nav up to the navbar
[#132108375](https://www.pivotaltracker.com/story/show/132108375)
